Bugfix: deleting "dir" symlinks

+++ src/core-filesystem/FileSystemPortable.cs
@@ -161,7 +161,7 @@ public class FileSystemPortable : IFileSystem {
 
     public void DeleteEntry(FileSystemEntry entry) {
       RemoveAccessDeniedAttributes(entry);
-      if (entry.IsDirectory) {
+      if (entry.IsDirectory && !entry.IsReparsePoint) {
         Directory.Delete(entry.Path.GetFullName(_fullNameBufferPool), recursive: false);
       } else {
         File.Delete(entry.Path.GetFullName(_fullNameBufferPool));
